requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
Django==1.11.20
django-cors-headers==3.0.1
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
idna==2.8
kiwisolver==1.1.0
matplotlib==2.2.4
mysql-connector==2.2.9
nltk==3.4.1
numpy==1.16.3
oauthlib==3.0.1
pkg-resources==0.0.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.4.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.3
textblob==0.15.3
tweepy==3.7.0
urllib3==1.24.3
